+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Added:
 - `--verbose` and `--debug` options set the log level
 - Option to set random seed, with updated tests
 - GP-2093 Bedtools coverage depth metrics
+- Additional metrics: Total coverage, coverage per target, coverage histogram
 
 v0.1.6: 2019-08-16
 ------------------